requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
numpy==1.19.5
opencv-python==4.5.3.56
Pillow==8.3.1
pydub==0.24.1
requests==2.26.0
Flask-Cors==3.0.10